requirements.txt
asyncpg
discord.py
discord
jishaku
ruamel.yaml